new solid sample trim

+++ element.h
@@ -3,12 +3,12 @@
 
 struct elementBase {
   int z;
-  float m, t; // mass and relative amount
+  double m, t; // mass and relative amount
 
-  float Edisp, Elbind; // displacement energy and lattice binding energy
+  double Edisp, Elbind; // displacement energy and lattice binding energy
 
   // calculated
-  float my, ec, ai, fi;
+  double my, ec, ai, fi;
 
   elementBase();
 };
